REFACTOR: took out the start class article and made it a reusable component
